Move NLF attribute components into NLF texts folder
