+++ client/anoixo-client/src/texts/NLF/Attributes/utils/TransliterationPossibilities.js
