+++ client/anoixo-client/src/texts/TextSettings.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import AttributeToComponentMap from '../query/Attributes/AttributeToComponentMap';
-import NLFAttributeToComponentMap from '../query/Attributes/NLF/NLFAttributeComponentMap';
+import NLFAttributeToComponentMap from './NLF/Attributes/NLFAttributeComponentMap';
 import { TextContext } from './TextContext';
 
 enum TextName {
